backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-generativeai
elevenlabs
stytch
apify-client
httpx
python-multipart